+++ src/boot/ksdk1.1.0/devBME680.h
@@ -44,7 +44,7 @@ WarpStatus	writeSensorRegisterBME680(uint8_t deviceRegister, uint8_t payload, ui
 WarpStatus	configureSensorBME680(	uint8_t payloadCtrl_Hum,
 					uint8_t payloadCtrl_Meas,
 					uint8_t payloadGas_0,
-					uint8_t menuI2cPullupValue);
+					uint16_t menuI2cPullupValue);
 WarpStatus	readSensorRegisterBME680(uint8_t deviceRegister, int numberOfBytes);
 WarpStatus	readSensorSignalBME680(WarpTypeMask signal,
 					WarpSignalPrecision precision,
