+++ src/boot/ksdk1.1.0/devL3GD20H.c
@@ -118,7 +118,7 @@ writeSensorRegisterL3GD20H(uint8_t deviceRegister, uint8_t payload, uint16_t men
 }
 
 WarpStatus
-configureSensorL3GD20H(uint8_t payloadCTRL1, uint8_t payloadCTRL2, uint8_t payloadCTRL5, uint8_t menuI2cPullupValue)
+configureSensorL3GD20H(uint8_t payloadCTRL1, uint8_t payloadCTRL2, uint8_t payloadCTRL5, uint16_t menuI2cPullupValue)
 {
 	WarpStatus	status1, status2, status3;
